# TypedApiTestProject/typedapitestproject.client/swagger/swagger.backup.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "TypedApiTestProject.Server",
    "version": "1.0"
  },
  "paths": {
...
    "/api/suppliers": {
      "get": {
        "tags": [
          "Supplier"
        ],
        "operationId": "GetSuppliers",
        "parameters": [
          {
            "name": "SupplierIds",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "format": "uuid"
              }
            }
          },
          {
            "name": "CompanyName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "CountryCodes",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "Verified",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "CreatedFrom",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "CreatedTo",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "PageNumber",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "SortBy",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "SortDirection",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/SortDirection"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierModelApiPaginationResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierModelApiPaginationResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierModelApiPaginationResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Supplier"
        ],
        "operationId": "CreateSupplier",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SupplierRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SupplierRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SupplierRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/suppliers/{id}": {
      "get": {
        "tags": [
          "Supplier"
        ],
        "operationId": "GetSupplierById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierModel"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Supplier"
        ],
        "operationId": "UpdateSupplier",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SupplierRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SupplierRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SupplierRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierModel"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Supplier"
        ],
        "operationId": "DeleteSupplier",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/suppliers/{id}/verify": {
      "post": {
        "tags": [
          "Supplier"
        ],
        "operationId": "VerifySupplier",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierModel"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierModel"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "SortDirection": {
        "enum": [
          "Default",
          "Neutral",
          "Asc",
          "Desc"
        ],
        "type": "string"
      },
      "SupplierModel": {
        "required": [
          "companyName",
          "contactEmail",
          "countryCode",
          "createdAt",
          "id",
          "verified"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "companyName": {
            "type": "string"
          },
          "contactEmail": {
            "type": "string"
          },
          "countryCode": {
            "type": "string"
          },
          "verified": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "SupplierModelApiPaginationResponse": {
        "required": [
          "data",
          "pageNumber",
          "pageSize",
          "totalCount",
          "totalPages",
          "totalRecords"
        ],
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SupplierModel"
            }
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "totalRecords": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "SupplierRequest": {
        "required": [
          "companyName",
          "contactEmail",
          "countryCode",
          "verified"
        ],
        "type": "object",
        "properties": {
          "companyName": {
            "type": "string"
          },
          "contactEmail": {
            "type": "string"
          },
          "countryCode": {
            "type": "string"
          },
          "verified": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
    }
  },
...
}